refactor: move getting hgvs loc start/end to HgvsTools
