+++ src/ga4gh/vrs/utils/hgvs_tools.py
@@ -5,8 +5,10 @@ import hgvs.dataproviders.uta
 import hgvs.normalizer
 import hgvs.variantmapper
 import hgvs.sequencevariant
+from hgvs.location import Interval, SimplePosition, AAPosition, BaseOffsetInterval
 
 import logging
+from typing import Optional, Union
 
 from ga4gh.vrs.dataproxy import _DataProxy, create_dataproxy
 from ga4gh.vrs import models
@@ -166,7 +168,44 @@ class HgvsTools():
 
         retval = {"refget_accession": refget_accession, "start": start, "end": end, "literal_sequence": state}
         return retval
-    
+
+    @staticmethod
+    def get_vrs_loc_start_end_val(
+        pos: Union[SimplePosition, AAPosition, BaseOffsetInterval, Interval],
+        is_start: bool = True
+    ) -> Union[int, models.Range]:
+        """Get VRS Location start or end value
+
+        :param pos: biocommons hgvs location instance for position
+        :param is_start: ``True`` if ``pos`` represents VRS Sequence Location start.
+            ``False`` if ``pos`` represents VRS Sequence Location end.
+        :raise ValueError: If unsupported biocommons hgvs location is passed
+        :return: VRS Location start or end value using inter-residue positions
+        """
+        def _get_pos_value(
+            position: Optional[int],
+            do_subtract_1: bool
+        ) -> Optional[int]:
+            """Get position value
+
+            :param position: Position
+            :param do_subtract_1: Whether or not we need to subtract 1 for ``position``
+            :return: Adjusted position value
+            """
+            return position - 1 if do_subtract_1 else position
+
+        if isinstance(pos, (SimplePosition, AAPosition, BaseOffsetInterval)):
+            vrs_loc_pos_val = _get_pos_value(pos.base, is_start)
+        elif isinstance(pos, Interval):
+            start_val = _get_pos_value(pos.start.base, is_start and pos.start.base is not None)
+            end_val = _get_pos_value(pos.end.base, is_start and pos.end.base is not None)
+            vrs_loc_pos_val = start_val if start_val == end_val else models.Range([start_val, end_val])
+        else:
+            err_msg = f"HGVS Location is not supported: {type(pos)}"
+            raise ValueError(err_msg)
+
+        return vrs_loc_pos_val
+
     def from_allele(self, vo, namespace=None):
         """generates a *list* of HGVS expressions for VRS Allele.
